+++ src/components/Collection/CollectionName.vue
@@ -24,6 +24,7 @@
           <el-input size="mini"
             v-model="collectionName"
             ref="collectionNameInput"
+            v-mask="'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'"
             style="width: 200px"
             @keyup.enter.native="handleEdit" />
           <el-button round
@@ -42,7 +43,7 @@
 
 <script>
 import VueTypes from 'vue-types';
-import DeletePopover from './DeletePopover';
+import DeletePopover from '@/components/UI/DeletePopover';
 
 export default {
   props: {