+++ src/components/UI/GoBackButton.vue
@@ -1,0 +1,4 @@
+<template>
+  <el-button type="text"
+    @click="$router.go(-1)">Go Back</el-button>
+</template>